fix: update type fiels in tbl_product

+++ database/migrations/2023_07_22_115356_create_products_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->float('price');
             $table->integer('quantity');
             $table->text('description')->nullable();
-            $table->tinyInteger('view')->default(0);
+            $table->integer('view')->default(0);
             $table->string('slug')->nullable();
             $table->integer('cate_id');
             $table->integer('brand_id');
